requirements.txt
Flask
pyTelegramBotAPI
python-dotenv
google-generativeai
gTTS
requests
elevenlabs
